+++ CHANGELOG.md
@@ -456,3 +456,4 @@
 - 2024-07-06: Docs mod, send, mod, mod, +6
 - 2024-07-07: Async user, site config, integration, embeddings, +2
 - 2024-07-08: Test mod, 0001 initial, site config, user, +6
+- 2024-07-09: Docs mod, mailbox, embeddings, user, +2